+++ public/js/PTL.main.js
@@ -357,18 +357,19 @@ var PTL = (function() {
 
       $fileImportInput.change(function(evt) {
         var f = evt.target.files[0],
-            existingFeeds = PTL.tab.list(),
+            existingFeeds = JSON.parse(JSON.stringify(PTL.tab.list())),
             reader = new FileReader();
 
         reader.onload = (function() {
           return function(e) {
 
             if (PTL.util.isValidJson(e.target.result)) {
-              PTL.tab.empty(function() {
-                PTL.tab.populate(JSON.parse(e.target.result));
-                // PTL.tab.populate(JSON.parse(existingFeeds.concat(e.target.result)), true);
-                // var finalObj = existingFeeds.concat(e.target.result);
-              });
+              PTL.dialog.importFeeds(existingFeeds, e.target.result);
+              // PTL.tab.empty(function() {
+              //   PTL.tab.populate(existingFeeds.concat(JSON.parse(e.target.result)));
+              //   // PTL.tab.populate(existingFeeds.concat(e.target.result));
+              //   // var finalObj = existingFeeds.concat(e.target.result);
+              // });
             } else if (PTL.util.isNV(e.target.result)) {
               PTL.util.importNV(e.target.result);
             } else {
@@ -401,7 +402,6 @@ var PTL = (function() {
             });
 
       // $debugHiddenButton.appendTo('body');
-
 
     },
     sideMenu: function(action) {